# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto API Specifications
  description: |-
    This OpenAPI document describes the WasaPhoto API.
  version: "1.0"
...
paths:
...
  /profiles/{username}:  
    get: 
      parameters:
      - name: username
        in: path
        required: true
        description: Username of the profile's owner the user wants to see
        schema:
          $ref: "#/components/schemas/username"
      - name: userId
        in: query
        required: true
        description: profile owner's userId
        schema:
          $ref: "#/components/schemas/userId"
      
      tags: ["View"]
      summary: View a profile
      description: |-
        This method allows a user to see in details the profile of another user. The method takes as input the username of the wanted profile and its link if found, an error otherwise.
      operationId: getUserProfile
      responses:
        "200": 
          description: "Profile succesfully retrieved"
          content:
            image/png:
              schema:
                description: array of photos of the requested profile
                type: array
                minItems: 0
                maxItems: 16
                example: [pngbinaryencoding1, pngbinaryencoding2]
                items:
                  $ref: '#/components/schemas/image'
        "404":
          description: User not found

        "403": 
          description: Unauthorized

        "500": 
          description: An internal error occurred
...
  /photos/:

    post:
      parameters:
      - name: userId
        description: profile's owner userId
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/userId'
      tags: ["Photos"]
      summary: Post a new photo 
      description: |-
          This method posts a new photo.
          A new photo is added even if there already exists an identical one.
          The method returns the photo Uri, if some problem during loading occurs an error is raised.
      operationId: uploadPhoto
      requestBody:
        description: Photo details
        content:
         image/png:
          schema:
            $ref: "#/components/schemas/image"
        required: true    

      responses:
        "201": 
          description: Photo succesfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/photoId'
        '400':
          description: The uploaded photo does not meet requirements 
        '403': 
          description: Unauthorized

        "404":
          description: User not found

        '500':
          description: An internal error occurred
  /photos/{photoId}:
    delete:
      tags: ["Photos"]
      summary: Delete a photo
      description: |
        This method allows user a who already follows user b to unfollow him/her. User a needs to specify in the request user b'username. The method returns user b' username and an error message in case of error.
      operationId: deletePhoto
      parameters:
        - name: photoId
          in: path
          required: true
          description: id of the photo the user wants to delete
          schema:
            $ref: '#/components/schemas/photoId'

        - name: userId
          in: query
          required: true
          description: owner's userId
          schema:
            $ref: '#/components/schemas/userId'

      responses:
        "204": 
          description: Photo succesfully deleted

        "403": 
          description: Unauthorized
          
        "404":
          description: Photo or user not found
 
        "500": 
          description: An internal error occurred
...
components:
  securitySchemes:
        bearerAuth:
          type: http
          scheme: bearer
          bearerFormat: JWT 
  schemas:
      username: 
        description: |
          Every user has a unique username he/she can use to log in, it is the name which users can use to find other users.
        type: string
        example: "User432"
        pattern: '^.*?$'
        minLength: 3
        maxLength: 16
      userId: 
        description: |
          It is the identifier the system uses to handle users
        type: string
        pattern: '^.*?$'
        minLength: 16
        maxLength: 16
        example: wu3v54u89ncf4n8o
...
      photoId:
        type: string
        description: photo unique identifier
        pattern: '^.*?$'
        minLength: 16
        maxLength: 16
        example: qcn3nu8f49q33cfg
      image:
        type: string
        format: binary
        description: image encoding
        pattern: '^.*?$'
        minLength: 1
        maxLength: 999
...
security:
  - bearerAuth: [] 